feat: expose a default Cargo binary, plus a library target that can be used from other Rust applications
